refactor(room): extract useRoomDerived from useRoomScreenState (C12)
New hook useRoomDerived computes seatViewModels, roleStats,
currentSchemaConstraints, and actionMessage via useMemo.
useRoomScreenState delegates to useRoomDerived after useRoomActions.
